Don’t cache.
Not caching makes it possible for you to change output size
and text, and not having to reload twice.
Yet changing the collage colour is still not updated.

It is updated on the server though. So apparently the swarmbot
script uses a cached representation of these images?

Also, Pango requires an integer image size

Also, I’m now using my given name in git

+++ SwarmGenerator/swarm_bot.py
@@ -29,7 +29,7 @@ def swarm_bot(output_image, images, text=None, square=True):
     if square == True:
         HEIGHT = 24 * scale
     else:
-        HEIGHT = 24 * 2 ** 0.5 * scale
+        HEIGHT = int(24 * 2 ** 0.5 * scale)
     bot.size(WIDTH, HEIGHT);
     bot.background(None)
